chore: move `output.rs` to the `types` folder
