+++ crates/rolldown_binding/src/types/mod.rs
@@ -1,1 +1,4 @@
+pub mod binding_output_asset;
+pub mod binding_output_chunk;
+pub mod binding_outputs;
 pub mod binding_rendered_module;
